+++ src/app/account/account.services.js
@@ -1,20 +1,17 @@
 const {
 	SUCCESS_MESSAGES: {
-		SIGN_UP_SUCCESSFULLY,
-		ACCOUNT_VERIFIED_SUCCESSFULLY,
-		LOGGED_OUT_SUCCESSFULLY,
 		DEACTIVATED_SUCCESSFULLY,
 		CHECK_MAIL_BOX,
 		ACTIVATED_SUCCESSFULLY,
+		ACCOUNT_DELETED_SUCCESSFULLY,
+		CANCELED_ACCOUNT_DELETION,
 	},
 	FAILIURE_MESSAGES: {
-		ACCOUNT_ALREADY_VERIFIED,
-		WRONG_EMAIL_OR_PASSWORD,
 		ACCOUNT_NEED_TO_BE_VERIFIED,
 		ACCOUNT_NEET_TO_BE_ACTIVE,
 		ACCOUNT_ALREADY_ACTIVE,
-		ACCOUNT_TEMP_DELETED,
 		ALREADY_HAVE_VALID_ACTIVATION_LINK,
+		ALREADY_CANCELED_ACCOUNT_DELETION,
 	},
 } = require("../../constants/messages");
 
@@ -29,71 +26,6 @@ const TokenHelper = require("./../../helpers/token");
 const HashHelper = require("./../../helpers/hash");
 
 class AccountServices {
-	static async signUp(payload) {
-		const { _id: accountId, role } = await AccountRepository.create({ ...payload });
-
-		const verificationToken = await TokenHelper.generateVerificationToken({
-			accountId,
-			role,
-		});
-
-		const verificationLink = `${process.env.BASE_URL}:${process.env.PORT}/auth/verify-email/${verificationToken}`;
-
-		// TODO: Send email
-		console.log({ verificationLink });
-
-		await AccountRepository.updateOne(accountId, { verificationToken });
-
-		return SIGN_UP_SUCCESSFULLY;
-	}
-
-	static async confirmVerification(token) {
-		const { accountId } = await TokenHelper.verifyVerificationToken(token);
-
-		const foundAccount = await AccountRepository.findOneById(accountId);
-
-		if (foundAccount && foundAccount.isVerified) {
-			throw new BadRequestException(ACCOUNT_ALREADY_VERIFIED);
-		}
-
-		await AccountRepository.updateOne(
-			accountId,
-			{ isVerified: true, isVerifiedAt: new Date() },
-			{ verificationToken: 1 }
-		);
-
-		return ACCOUNT_VERIFIED_SUCCESSFULLY;
-	}
-
-	static async logIn({ email, password }) {
-		const foundAccount = await AccountRepository.findOne(email);
-
-		if (!foundAccount) {
-			throw new UnAuthorizedException(WRONG_EMAIL_OR_PASSWORD);
-		}
-
-		AccountServices.isVerifiedActiveNotDeleted(foundAccount);
-
-		const isPasswordCorrect = await HashHelper.verify(password, foundAccount.password);
-
-		if (!isPasswordCorrect) {
-			throw new UnAuthorizedException(WRONG_EMAIL_OR_PASSWORD);
-		}
-
-		// TODO: what is 2fa are enalbed? refactor!
-
-		return await SessionServices.create({
-			accountId: foundAccount._id,
-			role: foundAccount.role,
-		});
-	}
-
-	static async logOut({ accountId, accessToken }) {
-		await SessionRepository.deleteOne({ accountId, accessToken });
-
-		return LOGGED_OUT_SUCCESSFULLY;
-	}
-
 	static async deactivate(accountId) {
 		await AccountRepository.updateOne(accountId, { isActive: false, activeStatusChangedAt: new Date() });
 
@@ -114,7 +46,7 @@ class AccountServices {
 		}
 
 		await AccountServices.isVerified(account.isVerified);
-		await AccountServices.isNotDeleted(account.isDeleted);
+
 		await AccountServices.#checkIfActivationTokenExists(account.activationToken);
 
 		const activationToken = await TokenHelper.generateActivationToken({
@@ -139,7 +71,6 @@ class AccountServices {
 
 		await AccountServices.#isAlreadyActive(account.isActive);
 		await AccountServices.isVerified(account.isVerified);
-		await AccountServices.isNotDeleted(account.isDeleted);
 
 		await AccountRepository.updateOne(
 			accountId,
@@ -150,11 +81,28 @@ class AccountServices {
 		return ACTIVATED_SUCCESSFULLY;
 	}
 
+	static async terminate(accountId) {
+		await AccountRepository.updateOne(accountId, { isDeleted: true, isDeletedAt: new Date() });
+
+		return ACCOUNT_DELETED_SUCCESSFULLY;
+	}
+
+	static async cancelTermination(accountId) {
+		const foundAccount = await AccountRepository.findOneById(accountId);
+
+		if (!foundAccount.isDeleted) {
+			throw new BadRequestException(ALREADY_CANCELED_ACCOUNT_DELETION);
+		}
+
+		await AccountRepository.updateOne(accountId, { isDeleted: false }, { isDeletedAt: 1 });
+
+		return CANCELED_ACCOUNT_DELETION;
+	}
+
 	// TODO: work more on temp delete and deleted
-	static isVerifiedActiveNotDeleted(account) {
+	static isVerifiedActive(account) {
 		AccountServices.isVerified(account.isVerified);
 		AccountServices.isActive(account.isActive);
-		AccountServices.isNotDeleted(account.isTempDeleted);
 	}
 
 	static isVerified(isVerified) {
@@ -166,12 +114,6 @@ class AccountServices {
 	static isActive(isActive) {
 		if (!isActive) {
 			throw new UnAuthorizedException(ACCOUNT_NEET_TO_BE_ACTIVE);
-		}
-	}
-
-	static isNotDeleted(isTempDeleted) {
-		if (isTempDeleted) {
-			throw new UnAuthorizedException(ACCOUNT_TEMP_DELETED);
 		}
 	}
