+++ src/app/password/password.routes.js
@@ -1,11 +1,11 @@
 const express = require("express");
 
 const { change, forget, reset } = require("./password.controllers");
-const { isAuthenticated, isVerified, isActive, isNotDeleted } = require("./../../middlewares/index");
+const { isAuthenticated, isVerified, isActive } = require("./../../middlewares/index");
 
 const router = express.Router();
 
-router.route("/change").put([isAuthenticated, isVerified, isActive, isNotDeleted], change);
+router.route("/change").put([isAuthenticated, isVerified, isActive], change);
 
 router.route("/forget").post(forget);
