+++ src/app/session/session.routes.js
@@ -1,14 +1,14 @@
 const express = require("express");
 
 const { findAll, cancel, renew } = require("./session.controllers");
-const { isAuthenticated, isVerified, isActive, isNotDeleted } = require("./../../middlewares/index");
+const { isAuthenticated, isVerified, isActive } = require("./../../middlewares/index");
 
 const router = express.Router();
 
 router
 	.route("/")
-	.get([isAuthenticated, isVerified, isActive, isNotDeleted], findAll)
-	.post([isAuthenticated, isVerified, isActive, isNotDeleted], cancel);
+	.get([isAuthenticated, isVerified, isActive], findAll)
+	.post([isAuthenticated, isVerified, isActive], cancel);
 
 router.route("/renew").post(renew);
