+++ src/middlewares/index.js
@@ -1,7 +1,6 @@
 const isAuthenticated = require("./isAuthenticated.middleware");
 const isVerified = require("./isVerified.middleware");
 const isActive = require("./isActive.middleware");
-const isNotDeleted = require("./isNotDeleted.middleware");
 
 // ** Need to be used in the same order!
 
@@ -9,5 +8,4 @@ module.exports = {
 	isAuthenticated,
 	isVerified,
 	isActive,
-	isNotDeleted,
 };
