--- src/middlewares/isNotDeleted.middleware.js
@@ -1,7 +1,0 @@
-const AccountServices = require("../app/account/account.services");
-
-module.exports = async (req, res, next) => {
-	AccountServices.isNotDeleted(req.isAccountDeleted);
-
-	next();
-};
